+++ pkg/config/config.go
@@ -1,4 +1,4 @@
-package utils
+package config
 
 import (
 	"gopkg.in/yaml.v3"
